Move BaseModel, Model and RestfulController
